Updated to include p_method parameter
send_ical_email now includes the method type. REQUEST = for new/updated meetings CANCEL = meeting cancellation. 

Prior to Exchange 2016 (and on most other mail servers) the method in the message body could be REQUEST for both new and cancelled meetings. Exchange 2016 will not correctly process the ical event content for cancellations if the method on the message body = REQUEST. The invite will delivered as a email with an attachment named "not supported calendar message.ics". I suspect that Exchange 2016 needs the method on the meesage body to be the same as the method on the ical event.